--- products/jbrowse-desktop/src/JBrowse.js
@@ -1,60 +1,0 @@
-import React, { useEffect, useState, Suspense } from 'react'
-import { observer } from 'mobx-react'
-import { CssBaseline, ThemeProvider } from '@material-ui/core'
-import { getConf } from '@jbrowse/core/configuration'
-import { App, createJBrowseTheme } from '@jbrowse/core/ui'
-import { AssemblyManager } from '@jbrowse/plugin-data-management'
-import StartScreen from './StartScreen'
-import factoryReset from './factoryReset'
-
-const JBrowse = observer(({ pluginManager }) => {
-  const [firstLoad, setFirstLoad] = useState(true)
-  const { rootModel } = pluginManager
-  const {
-    session,
-    jbrowse,
-    error,
-    isAssemblyEditing,
-    setAssemblyEditing,
-  } = rootModel
-
-  useEffect(() => {
-    if (firstLoad && session) {
-      setFirstLoad(false)
-    }
-  }, [firstLoad, session])
-
-  if (error) {
-    throw new Error(error)
-  }
-
-  const theme = getConf(jbrowse, 'theme')
-
-  return (
-    <ThemeProvider theme={createJBrowseTheme(theme)}>
-      <CssBaseline />
-      {session ? (
-        <>
-          <App session={session} />
-          <Suspense fallback={<div />}>
-            <AssemblyManager
-              rootModel={rootModel}
-              open={isAssemblyEditing}
-              onClose={() => {
-                setAssemblyEditing(false)
-              }}
-            />
-          </Suspense>
-        </>
-      ) : (
-        <StartScreen
-          rootModel={rootModel}
-          bypass={firstLoad}
-          onFactoryReset={factoryReset}
-        />
-      )}
-    </ThemeProvider>
-  )
-})
-
-export default JBrowse
