+++ products/jbrowse-desktop/src/StartScreen/RecentSessionsPanel.tsx
