+++ products/jbrowse-web/src/Loader.tsx
@@ -744,7 +744,10 @@ const Renderer = observer(
       if (!pm.rootModel?.session) {
         return (
           <Suspense fallback={<div>Loading...</div>}>
-            <StartScreen root={pm.rootModel} onFactoryReset={factoryReset} />
+            <StartScreen
+              rootModel={pm.rootModel}
+              onFactoryReset={factoryReset}
+            />
           </Suspense>
         )
       }
